Add files via upload
This is my first IoT project working with LoRa devices. The idea behind this project is to have a mobile LoRa node(LoPy with pymkr) that can be placed in Pets neck and the owner of the pet will be able to send command, see the distance of the pet using another LoPy device. This small project is done using two LoPy modules and setting up point to point transmission. Based on the RSSI value of the signal master device will be able to calculate approximate distance and also send some commands to the Pet. It is a group project done for the subject IoT offered at Uni Bremen. Coding was done in MicroPython.